+++ Runtime/AssemblyInfo.cs
@@ -1,3 +1,5 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Appstack.Unity.Tests")]
+[assembly: InternalsVisibleTo("Appstack.Unity.Editor")]
+[assembly: UnityEngine.Scripting.AlwaysLinkAssembly]